requirements.txt
Flask>=2.0
requests
reportlab
Pillow
flask
fpdf
PyGithub
